rdtsc can be issued as a normal user, so that is ok.
rdmsr is more complicated: it needs to be ring 0, so it could only run in a kernel module. I will decide later if I want to do that.

As a workaround, I am using the rdmsr kernel module to read msr values. Comparing issueing rdtsc directly and rdmsr for 0x10 give the same result, which is promising.